Add removeAll thunk
