+++ src/features/cart/addToCart/ui/RemoveIcon/RemoveIcon.tsx
@@ -1,0 +1,35 @@
+import { useCallback } from 'react'
+import { type ProductId } from '@/entities/product'
+import { removeAllProductFromCartThunk } from '@/features/cart/addToCart/model/addProductToCart'
+import { useConfirmModal } from '@/shared/lib'
+import { useAppDispatch } from '@/shared/model'
+import { Icon } from '@/shared/ui'
+
+type Props = {
+  productId: ProductId
+}
+
+export function RemoveIcon(props: Props) {
+  const dispatch = useAppDispatch()
+  const confirmRemoveModal = useConfirmModal()
+
+  const onClickToRemove = useCallback(
+    (event: React.MouseEvent<HTMLDivElement>) => {
+      event.preventDefault()
+
+      confirmRemoveModal.show({
+        title: 'Are you really want remove product from cart?',
+        confirmText: 'Yes',
+        cancelText: 'No',
+        onConfirm: () => {
+          confirmRemoveModal.remove()
+          dispatch(removeAllProductFromCartThunk(props.productId))
+        },
+        onCancel: () => confirmRemoveModal.remove(),
+      })
+    },
+    [props.productId]
+  )
+
+  return <Icon type="trash" onClick={onClickToRemove} />
+}
